Refactor storage engine around core and broadcast modules
