+++ apps/web/src/lib/stores/storage/README.md
@@ -16,7 +16,9 @@ Svelte stores around a serialised snapshot. The goals are:
 | [`types.ts`](./types.ts)         | Declares the snapshot schema (installation meta, configuration, settings, document index, history, audit trails).                                                                                                                                                           |
 | [`defaults.ts`](./defaults.ts)   | Builds default runtime configuration, UI settings, and the initial snapshot that seeds a new installation.                                                                                                                                                                  |
 | [`driver.ts`](./driver.ts)       | Wraps the host persistence API. The default driver uses `localStorage` but any alternative (memory, IndexedDB, remote APIs) can be swapped in. The driver also exposes a `subscribe` hook so the engine can react to external writes (e.g. storage events from other tabs). |
-| [`engine.ts`](./engine.ts)       | Turns snapshots into writable Svelte stores (`snapshot`, `config`, `settings`) and exposes lifecycle helpers (`refresh`, `reset`, `update`). The engine coordinates with the driver, applying mutations and keeping derived stores in sync.                                 |
+| [`broadcast.ts`](./broadcast.ts) | Coalesces outbound snapshot notifications, preferring `BroadcastChannel` and falling back to emitting synthetic `storage` events when the channel API is unavailable.                                                                                                       |
+| [`core.ts`](./core.ts)           | Owns snapshot orchestration: hydration, persistence, history management, and broadcast coordination through a framework-agnostic observable API.                                                                                                                            |
+| [`engine.ts`](./engine.ts)       | Adapts the core to Svelte by wrapping the observable API in writable stores (`snapshot`, `config`, `settings`) and forwarding lifecycle helpers (`refresh`, `reset`, `update`).                                                                                             |
 | [`index.ts`](./index.ts)         | Convenience barrel export for the storage subsystem.                                                                                                                                                                                                                        |
 
 ## Data flow at runtime
