+++ apps/web/src/lib/stores/storage/index.ts
@@ -1,5 +1,7 @@
 export * from "./constants";
 export * from "./defaults";
+export * from "./broadcast";
+export * from "./core";
 export * from "./driver";
 export * from "./engine";
 export * from "./history";
