+++ specs/storage.spec.md
@@ -212,7 +212,7 @@ These scenarios serve both as documentation and as test case inspiration:
 
 ## 15. Storage Engine Contract
 
-The storage engine orchestrates persistence and read models for the rest of the app.
+The storage runtime is split into a framework-agnostic **Storage Core** that orchestrates persistence/read models and a Svelte-facing engine that adapts the core into readable stores. Callers interact with the engine, while tests can exercise the core directly without touching UI concerns.
 
 - **Initialization**
   - On boot, the engine loads the latest snapshot from the driver; if none is present it seeds defaults generated by `createInitialSnapshot`.
@@ -234,7 +234,7 @@ The storage engine orchestrates persistence and read models for the rest of the 
   - After every successful update, the `config` and `settings` stores are synchronized with the latest snapshot fields.
 
 - **Broadcasts**
-  - `scheduleBroadcast` is currently a stub that preserves the contract surface; future work will hook this into `BroadcastChannel` (or similar).
+  - `scheduleBroadcast` coalesces messages until the next macrotask, prefers `BroadcastChannel` when available, and falls back to emitting synthetic `storage` events when the channel API fails or is unavailable.
 
 ## 16. Snapshot Schema & Invariants
 
@@ -671,7 +671,8 @@ This section is for the AI or developers to update after implementation runs.
   - Privacy toggles, encryption hooks, and compliance-aware audit policies from §25.
 
 - **Test files and coverage**:
-  - Unit tests for storage engine hydration, refresh, reset, and update semantics: `/apps/web/src/lib/stores/storage/engine.test.ts`
+  - Unit tests for storage core hydration, refresh, reset, and update semantics: `/apps/web/src/lib/stores/storage/core.test.ts`
+  - Integration tests for the Svelte engine wrapper: `/apps/web/src/lib/stores/storage/engine.test.ts`
   - Unit tests for history capture and eviction: `/apps/web/src/lib/stores/storage-history.test.ts`
   - Unit tests for migration upgrades and corruption fallback: `/apps/web/src/lib/stores/storage-migration.test.ts`
   - Unit tests for multi-tab synchronisation and broadcast throttling: `/apps/web/src/lib/stores/storage-sync.test.ts`
